try to fix posix path issues
