+++ packages/docusaurus-utils/src/__tests__/posixPath.test.ts
@@ -1,0 +1,27 @@
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+import {posixPath} from '../posixPath';
+
+describe('posixPath', () => {
+  test('posixPath works', () => {
+    const asserts: Record<string, string> = {
+      'c:/aaaa\\bbbb': 'c:/aaaa/bbbb',
+      'c:\\aaaa\\bbbb\\★': 'c:\\aaaa\\bbbb\\★',
+      'D:\\a\\docusaurus\\docusaurus\\website\\static\\dogfooding\\新控制器空间\\图片.png':
+        'D:/a/docusaurus/docusaurus/website/static/dogfooding/新控制器空间/图片.png',
+      '\\\\?\\c:\\aaaa\\bbbb': '\\\\?\\c:\\aaaa\\bbbb',
+      'c:\\aaaa\\bbbb': 'c:/aaaa/bbbb',
+      'foo\\bar': 'foo/bar',
+      'foo\\bar/lol': 'foo/bar/lol',
+      'website\\docs/**/*.{md,mdx}': 'website/docs/**/*.{md,mdx}',
+    };
+    Object.keys(asserts).forEach((file) => {
+      expect(posixPath(file)).toBe(asserts[file]);
+    });
+  });
+});
